rename feature to derive
