+++ rust/opendp_derive/Cargo.toml
@@ -21,5 +21,5 @@ quote = { workspace = true, optional = true }
 opendp_tooling = { path = "../opendp_tooling", optional = true  }
 
 [features]
-full = ["syn", "quote", "opendp_tooling"]
+derive = ["syn", "quote", "opendp_tooling"]
 local-docs = []
